moved ellipse functions into their own category
