+++ CircleTheVortex/aggregation/vortex_cluster.py
@@ -1,176 +1,9 @@
 from .utils import pixel_to_lonlat, lonlat_to_pixel
+from .shape_funcs import average_shape_IoU, IoU_metric, \
+    params_to_shape
 import numpy as np
-import scipy.optimize
 from copy import deepcopy
 import tqdm
-from shapely.geometry import Point, Polygon
-from shapely import affinity
-
-
-def scale_shape(params, gamma):
-    '''
-        Scales the ellipse by a factor of gamma
-    '''
-    return [
-        # center is the same
-        params[0],
-        params[1],
-        # radius_x and radius_y scale
-        gamma * params[2],
-        gamma * params[3],
-        # angle does not change
-        params[4]
-    ]
-
-
-def get_sigma_shape(params, sigma):
-    '''
-        Return the plus and minus one sigma shape given the starting parameters
-        and sigma value.
-        Parameters
-        ----------
-        params : list
-            A list of the parameters for the shape (as defined by PFE)
-        shape : string
-            The name of the shape these parameters belong to
-            (see :meth:`panoptes_to_geometry` for supported shapes)
-        sigma : float
-            The standard deviation used to scale up and down the input shape
-        Returns
-        -------
-        plus_sigma : list
-            A list of shape parameters for the 1-sigma scaled up average
-        minus_sigma : list
-            A list of shape parameters for the 1-sigma scaled down average
-    '''
-    gamma = np.sqrt(1 - sigma)
-    plus_sigma = scale_shape(params, 1 / gamma)
-    minus_sigma = scale_shape(params, gamma)
-    return plus_sigma, minus_sigma
-
-
-def params_to_shape(params):
-    '''
-        Converts the parameter list for an ellipse
-        to a `shapely.geometry.Polygon` object
-    '''
-    x, y, rx, ry, angle = params
-    circ = Point((x, y)).buffer(1)
-    ell = affinity.scale(circ, rx, ry)
-    ellr = affinity.rotate(ell, -angle)
-
-    return ellr
-
-
-def average_bounds(params_list):
-    '''Find the bounding box for the average shape for each of the shapes
-    parameters.
-    Parameters
-    ----------
-    params_list : list
-        A list of shape parameters that are being averaged
-    shape : string
-        The shape these parameters belong to
-        (see :meth:`panoptes_to_geometry` for supported shapes)
-    Returns
-    -------
-    bound : list
-        This is a list of tuples giving the min and max bounds for
-        each shape parameter.
-    '''
-    geo = params_to_shape(params_list[0])
-    # Use the union of all shapes to find the bounding box
-    for params in params_list[1:]:
-        geo = geo.union(params_to_shape(params))
-    # bound on x
-    bx = (geo.bounds[0], geo.bounds[2])
-    # bound on y
-    by = (geo.bounds[1], geo.bounds[3])
-    # width of geo
-    dx = bx[1] - bx[0]
-    # height of geo
-    dy = by[1] - by[0]
-    # bound is a list of tuples giving (min, max) values
-    # for each paramters of the shape
-    bound = [bx, by]
-    # bound on width or radius_x, min set to 1 pixel
-    bound.append((1, dx))
-    # bound on height or radius_y, min set to 1 pixel
-    bound.append((1, dy))
-    # bound on angle (capped at 180 due to symmetry)
-    bound.append((0, 180))
-
-    return bound
-
-
-def average_shape_IoU(params_list, probs, shape='ellipse'):
-    '''Find the average shape and standard deviation from a
-    list of parameters with respect to the IoU metric.
-    Parameters
-    ----------
-    params_list : list
-        A list of shape parameters that are being averaged
-    shape : string
-        The shape these parameters belong to
-        (see :meth:`panoptes_to_geometry` for supported shapes)
-    Returns
-    -------
-    average_shape : list
-        A list of shape parameters for the average shape
-    sigma : float
-        The standard deviation of the input shapes with
-        respect to the IoU metric
-    '''
-    def sum_distance(x):
-        return sum([probs[i] * IoU_metric(params_to_shape(x),
-                                          params_to_shape(params_list[i]),
-                                          reshape=False)**2
-                    for i in range(len(params_list))])
-    # find shape that minimizes the variance in the IoU metric using bounds
-    m = scipy.optimize.shgo(
-        sum_distance,
-        sampling_method='sobol',
-        bounds=average_bounds(params_list)
-    )
-    # find the 1-sigma value
-    sigma = np.sqrt(m.fun / (len(params_list) - 1))
-    return list(m.x), sigma
-
-
-def IoU_metric(params1, params2, reshape=True):
-    '''Find the Intersection of Union distance between two shapes.
-    Parameters
-    ----------
-    params1 : list
-        A list of the parameters for shape 1 (as defined by PFE)
-    params2 : list
-        A list of the parameters for shape 2 (as defined by PFE)
-    shape : string
-        The shape these parameters belong to
-        (see :meth:`panoptes_to_geometry` for supported shapes)
-    Returns
-    -------
-    distance : float
-        The IoU distance between the two shapes.
-        0 means the shapes are the same,
-        1 means the shapes don't overlap, values in the middle mean partial
-        overlap.
-    '''
-    if reshape:
-        par1 = params1.reshape((65, 2))
-        par2 = params2.reshape((65, 2))
-    else:
-        par1 = params1
-        par2 = params2
-
-    geo1 = Polygon(par1)
-    geo2 = Polygon(par2)
-    intersection = geo1.intersection(geo2).area
-    union = geo1.union(geo2).area
-    if union == 0:
-        # catch divide by zero (i.e. cases when neither shape has an area)
-        return np.inf
-    return 1 - intersection / union
 
 
 def cluster_vortices(ellipses):
@@ -205,13 +38,19 @@ def cluster_vortices(ellipses):
 
             delete_mask = np.where(IoUs > 0)[0]
 
+            # check the IoUs
             if IoUs[1:].sum() == 0:
+                # if there are no overlapping vortices
+                # then we add this to the lone vortex bin
                 lone_ellipses.append(elli)
             else:
+                # if not, then we add all overlapping 
+                # vortices to the group
                 ellipse_groups.append(ellipse_queue[delete_mask])
 
             pbar.update(len(delete_mask))
 
+            # remove the overlapping elements from the group
             ellipse_queue = np.delete(ellipse_queue, delete_mask)
 
     print(f"{len(lone_ellipses)} vortices; {len(ellipse_groups)} groups")
